Separate skills editor to its own file
